Added keyboard support Added getElementsByTag Changed OpenGL calls glBegin/glEnd to glVertexPointer/glDrawArrays

+++ src/size.h
@@ -9,10 +9,10 @@ class Size
 	public:
 		Size();
 		
-		inline void setWidth(int width) { m_width = width; m_widthAuto = false; }
-		inline void setHeight(int height) { m_height = height; m_heightAuto = false; }
-		inline int getWidth() const { return m_width; }
-		inline int getHeight() const { return m_height; }
+		inline void setWidth(float width) { m_width = width; m_widthAuto = false; }
+		inline void setHeight(float height) { m_height = height; m_heightAuto = false; }
+		inline float getWidth() const { return m_width; }
+		inline float getHeight() const { return m_height; }
 		
 		inline void setWidthAuto() { m_widthAuto = true; }
 		inline void setHeightAuto() { m_heightAuto = true; }
@@ -20,8 +20,8 @@ class Size
 		inline bool getHeightAuto() const { return m_heightAuto; }
 		
 	private:
-		int m_width;
-		int m_height;
+		float m_width;
+		float m_height;
 		bool m_widthAuto;
 		bool m_heightAuto;
 };
